+++ Jenkinsfile-sts.groovy
@@ -353,7 +353,7 @@ pipeline {
                 steps { 
                     script   {
                          // node('docker-slave-nodebackend-buster-12.x') {  
-                        stage('Debian Build') {
+                        stage('Build Debian Folder') {
                             try {                         
                                 echo "Build debian pkg ${params.RAINBOWNODESDKVERSION} ${workspace}"
                                 sh script: """
@@ -383,10 +383,28 @@ pipeline {
                                 sed "s/ref:doc\\/sdk\\/node\\//ref:doc\\/sdk\\/node\\/sts\\//g" "index.yml"  |tee "Documentation/doc/sdk/node/sts/index.yml"                      
                                 sed "s/\\/doc\\/sdk\\/node\\//\\/doc\\/sdk\\/node\\/sts\\//g" "sitemap.xml"  |tee "Documentation/doc/sdk/node/sts/sitemap.xml"                      
                                  
-                                #find Documentation/
-                                #cd "${workspace}/Documentation"
                                 """
+                            } catch (Exception e) {
+                                echo "Failure: ${currentBuild.result}: ${e}"
+                            }
+                        }
+                          
+                        stage("Generate documentation search index") {
+                            try {
+                                generateHubV2DocumentationSearchIndex("Documentation/doc/sdk/node/sts")
+                            } catch (Exception e) {
+                                echo "Failure: ${currentBuild.result}: ${e}"
+                            }
+                        }
+
+                        stage('Build Debian package') {
+                            try {
                                 echo "Build debian the package : "
+                                sh script: """
+                                    #find Documentation/
+                                    #cd "${workspace}/Documentation"
+                                """
+                                
                                 debianBuild(
                                     debianPath: 'Documentation',
                                     nextVersion: "${params.RAINBOWNODESDKVERSION}" ,
@@ -394,6 +412,9 @@ pipeline {
                                 )
                             } catch (Exception e) {
                                 echo "Failure: ${currentBuild.result}: ${e}"
+                            }
+                            finally {
+                                //    notifyBuild(currentBuild.result)
                             }
                         }